code cleanup, new minute scaling, textures rearrangement
